@@ -1,15 +1,14 @@
 ﻿using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
 
-namespace MongoDriver.Facade
+namespace MongoDriver.Facade;
+
+public class MongoDbCollection
 {
-    public class MongoDbCollection
-    {
-        /// <summary>
-        /// Represent _id field of collection.
-        /// </summary>
-        [BsonId]
-        [BsonRepresentation(BsonType.ObjectId)]
-        public virtual string Id { get; set; }
-    }
+    /// <summary>
+    /// Represent _id field of collection.
+    /// </summary>
+    [BsonId]
+    [BsonRepresentation(BsonType.ObjectId)]
+    public virtual string Id { get; set; }
 }
