@@ -1,22 +1,21 @@
 ﻿using MongoDB.Driver;
 
-namespace MongoDriver.Facade
+namespace MongoDriver.Facade;
+
+/// <summary>
+/// Represents extensions methods on MongoDbContext.
+/// </summary>
+public static class MongoDbContextExtensions
 {
     /// <summary>
-    /// Represents extensions methods on MongoDbContext.
+    /// Extension method which return collection.
+    /// This can be used to query.
     /// </summary>
-    public static class MongoDbContextExtensions
+    /// <typeparam name="TCollection">Collection model.</typeparam>
+    /// <param name="mongoDbContext"></param>
+    /// <returns>Generic IMongoCollection of collection type.</returns>
+    public static IMongoCollection<TCollection> Set<TCollection>(this MongoDbContext mongoDbContext) where TCollection : MongoDbCollection
     {
-        /// <summary>
-        /// Extension method which return collection.
-        /// This can be used to query.
-        /// </summary>
-        /// <typeparam name="TCollection">Collection model.</typeparam>
-        /// <param name="mongoDbContext"></param>
-        /// <returns>Generic IMongoCollection of collection type.</returns>
-        public static IMongoCollection<TCollection> Set<TCollection>(this MongoDbContext mongoDbContext) where TCollection : MongoDbCollection
-        {
-            return mongoDbContext.Database.GetCollection<TCollection>(typeof(TCollection).Name);
-        }
+        return mongoDbContext.Database.GetCollection<TCollection>(typeof(TCollection).Name);
     }
 }
