@@ -1,11 +1,10 @@
-﻿namespace MongoDriver.Facade.TestApi.Models
+﻿namespace MongoDriver.Facade.TestApi.Models;
+
+public class Location : MongoDbCollection
 {
-    public class Location : MongoDbCollection
-    {
-        public string Country { get; set; }
+    public string Country { get; set; }
 
-        public string City { get; set; }
+    public string City { get; set; }
 
-        public string Stree { get; set; }
-    }
+    public string Stree { get; set; }
 }
