@@ -5,42 +5,41 @@ using MongoDriver.Facade.TestApi.Models;
 using System.Collections.Generic;
 using Xunit;
 
-namespace Test
+namespace Test;
+
+public class ReflectionHelperTest
 {
-    public class ReflectionHelperTest
+    [Fact]
+    public void GetCollectionTest()
     {
-        [Fact]
-        public void GetCollectionTest()
-        {
-            // Arrange
-            var expected = new Dictionary<string, CreateCollectionOptions>();
+        // Arrange
+        var expected = new Dictionary<string, CreateCollectionOptions>();
 
-            // Book
-            var bookConfigration = new BookConfigration().Configure();
-            expected.Add(nameof(Book), bookConfigration);
+        // Book
+        var bookConfigration = new BookConfigration().Configure();
+        expected.Add(nameof(Book), bookConfigration);
 
-            // Food
-            expected.Add(nameof(Food), null);
+        // Food
+        expected.Add(nameof(Food), null);
 
-            // Location
-            expected.Add(nameof(Location), null);
+        // Location
+        expected.Add(nameof(Location), null);
 
-            // Person
-            expected.Add(nameof(Person), null);
+        // Person
+        expected.Add(nameof(Person), null);
 
-            // Post
-            expected.Add(nameof(Post), null);
+        // Post
+        expected.Add(nameof(Post), null);
 
-            // Product
-            var productConfigration = new ProductConfigration().Configure();
-            expected.Add(nameof(Product), productConfigration);
+        // Product
+        var productConfigration = new ProductConfigration().Configure();
+        expected.Add(nameof(Product), productConfigration);
 
-            // Act
-            var actual = new ReflectionHelper().GetCollections();
+        // Act
+        var actual = new ReflectionHelper().GetCollections();
 
-            // Assert
-            actual.Should().BeEquivalentTo(expected);
+        // Assert
+        actual.Should().BeEquivalentTo(expected);
 
-        }
     }
 }
